refactor(api-contracts): extract catalog endpoint resolver module (Refs #568)
Address dual-axis review findings for docs endpoint resolver:
- Split CatalogEndpoint/to_request resolution into catalog_endpoints.py
  so rust_source.py stays under the 1k-line maintainability bar
- Share match-self arm walking and Enum::Variant suffix parsing
- Collapse multi-path to_request alias expansion into one candidate pipeline
- Drop dead whitespace assignment in extract_to_request_receiver
- Document BaikeApiV1 independent enum break under CHANGELOG Unreleased
